requirements.txt
numpy==1.19.5
tensorflow==2.4.1
Flask==1.1.2
Pillow==9.1.1
